# docs/api/swagger.yaml
openapi: 3.0.3
info:
  title: AIPC Platform API
  description: 'AIPC Platform REST API for edge AI computing platform.


    Provides unified access to:

    - AI Runtime (model management, inference)

    - Event Bus (pub/sub messaging)

    - Device Control (camera, PTZ, GPIO)

    - App Manager (container lifecycle)

    - Stream (HLS video streaming)

    '
  version: 1.0.0
  contact:
    name: AIPC Platform Team
    email: support@camthink.ai
servers:
- url: http://localhost:8080/api/v1
  description: Development server
- url: http://{host}:8080/api/v1
  description: Production server
  variables:
    host:
      default: localhost
security:
- BearerAuth: []
...
paths:
...
  /apps/{app_id}/logs:
    get:
      tags:
      - App Manager
      summary: Get application logs
      operationId: getAppLogs
      parameters:
      - name: app_id
        in: path
        required: true
        schema:
          type: string
      - name: max_lines
        in: query
        schema:
          type: integer
          default: 100
      - name: follow
        in: query
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: Application logs
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/APIResponse'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        type: string
...
  /containers/{id}/logs:
    get:
      tags: [Containers]
      summary: Get container logs
      operationId: getContainerLogs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: tail
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Container logs
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/APIResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        logs:
                          type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
  schemas:
    APIResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: Success
        data:
          type: object
          description: Response payload
        error:
          $ref: '#/components/schemas/ErrorDetail'
...
    ErrorDetail:
      type: object
      properties:
        type:
          type: string
        detail:
          type: string

    # ==================== Store Schemas ====================
...